Department as FC, logic divided better :)
