+++ src/components/Application-Body/JobsAppBody.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import Button from '@material-ui/core/Button';
 import makeStyles from '@material-ui/core/styles/makeStyles';
 
-import PostNewJob from './PostNewJob';
+import PostNewJob from './PostNewJob/PostNewJob';
 
 const dialogThemeColor = "rgb(89,89,89)";
 const styles = makeStyles({
