+++ src/components/Application-Header/JobsAppBar.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+
 import { withStyles, WithStyles } from '@material-ui/core';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
